# Support fast path for Spark comparison function in join filter