--- a/velox/functions/lib/SIMDComparisonUtil.h
+++ b/velox/functions/lib/SIMDComparisonUtil.h
@@ -16,6 +16,7 @@
 #pragma once
 
 #include "velox/expression/VectorFunction.h"
+#include "velox/vector/DictionaryVector.h"
 
 namespace facebook::velox::functions {
 
@@ -67,11 +68,13 @@ inline uint64_t to64Bits(const int8_t* resultData) {
 
 template <typename A, typename B, typename Compare>
 void applyAutoSimdComparisonInternal(
+    Compare cmp,
+    VectorPtr& result,
     const SelectivityVector& rows,
     const A* __restrict rawA,
     const B* __restrict rawB,
-    Compare cmp,
-    VectorPtr& result) {
+    const vector_size_t* __restrict indexA,
+    const vector_size_t* __restrict indexB) {
   int8_t tempBuffer[64];
   int8_t* __restrict resultData = tempBuffer;
   const vector_size_t rowsBegin = rows.begin();
@@ -83,12 +86,12 @@ void applyAutoSimdComparisonInternal(
     auto i = 0;
     for (; i + 64 <= rowsEnd; i += 64) {
       for (auto j = 0; j < 64; ++j) {
-        resultData[j] = cmp(rawA, rawB, i + j) ? -1 : 0;
+        resultData[j] = cmp(rawA, rawB, indexA, indexB, i + j) ? -1 : 0;
       }
       rawResult[i / 64] = to64Bits(resultData);
     }
     for (; i < rowsEnd; ++i) {
-      bits::setBit(rawResult, i, cmp(rawA, rawB, i));
+      bits::setBit(rawResult, i, cmp(rawA, rawB, indexA, indexB, i));
     }
   } else {
     static constexpr uint64_t kAllSet = -1ULL;
@@ -103,7 +106,8 @@ void applyAutoSimdComparisonInternal(
           const size_t start = idx * 64;
           while (word) {
             auto index = start + __builtin_ctzll(word);
-            bits::setBit(rawResult, index, cmp(rawA, rawB, index));
+            bits::setBit(
+                rawResult, index, cmp(rawA, rawB, indexA, indexB, index));
             word &= word - 1;
           }
         },
@@ -113,13 +117,15 @@ void applyAutoSimdComparisonInternal(
           if (kAllSet == word) {
             // Do 64 comparisons in a batch, set results by SIMD.
             for (size_t row = 0; row < 64; ++row) {
-              resultData[row] = cmp(rawA, rawB, row + start) ? -1 : 0;
+              resultData[row] =
+                  cmp(rawA, rawB, indexA, indexB, row + start) ? -1 : 0;
             }
             rawResult[idx] = to64Bits(resultData);
           } else {
             while (word) {
               auto index = __builtin_ctzll(word);
-              resultData[index] = cmp(rawA, rawB, start + index) ? -1 : 0;
+              resultData[index] =
+                  cmp(rawA, rawB, indexA, indexB, start + index) ? -1 : 0;
               word &= word - 1;
             }
             // Set results only for selected rows.
@@ -237,70 +243,189 @@ void applyAutoSimdComparison(
     const B* __restrict rawB =
         args[1]->asUnchecked<FlatVector<B>>()->template rawValues<B>();
     detail::applyAutoSimdComparisonInternal(
-        rows,
-        rawA,
-        rawB,
-        [&](const A* __restrict rawA, const B* __restrict rawB, int i) {
+        [&](const A* __restrict rawA,
+            const B* __restrict rawB,
+            const vector_size_t* __restrict indexA,
+            const vector_size_t* __restrict indexB,
+            int i) {
           if constexpr (sizeof...(cmpArgs) > 0) {
             return Compare::apply(rawA[i], rawB[i], cmpArgs...);
           } else {
             return cmp(rawA[i], rawB[i]);
           }
         },
-        result);
+        result,
+        rows,
+        rawA,
+        rawB,
+        nullptr,
+        nullptr);
   } else if (args[0]->isConstantEncoding() && args[1]->isFlatEncoding()) {
     const A constA = args[0]->asUnchecked<ConstantVector<A>>()->valueAt(0);
     const A* __restrict rawA = &constA;
     const B* __restrict rawB =
         args[1]->asUnchecked<FlatVector<B>>()->template rawValues<B>();
     detail::applyAutoSimdComparisonInternal(
-        rows,
-        rawA,
-        rawB,
-        [&](const A* __restrict rawA, const B* __restrict rawB, int i) {
+        [&](const A* __restrict rawA,
+            const B* __restrict rawB,
+            const vector_size_t* __restrict indexA,
+            const vector_size_t* __restrict indexB,
+            int i) {
           if constexpr (sizeof...(cmpArgs) > 0) {
             return Compare::apply(rawA[0], rawB[i], cmpArgs...);
           } else {
             return cmp(rawA[0], rawB[i]);
           }
         },
-        result);
+        result,
+        rows,
+        rawA,
+        rawB,
+        nullptr,
+        nullptr);
   } else if (args[0]->isFlatEncoding() && args[1]->isConstantEncoding()) {
     const A* __restrict rawA =
         args[0]->asUnchecked<FlatVector<A>>()->template rawValues<A>();
     const B constB = args[1]->asUnchecked<ConstantVector<B>>()->valueAt(0);
     const B* __restrict rawB = &constB;
     detail::applyAutoSimdComparisonInternal(
-        rows,
-        rawA,
-        rawB,
-        [&](const A* __restrict rawA, const B* __restrict rawB, int i) {
+        [&](const A* __restrict rawA,
+            const B* __restrict rawB,
+            const vector_size_t* __restrict indexA,
+            const vector_size_t* __restrict indexB,
+            int i) {
           if constexpr (sizeof...(cmpArgs) > 0) {
             return Compare::apply(rawA[i], rawB[0], cmpArgs...);
           } else {
             return cmp(rawA[i], rawB[0]);
           }
         },
-        result);
+        result,
+        rows,
+        rawA,
+        rawB,
+        nullptr,
+        nullptr);
   } else if (args[0]->isConstantEncoding() && args[1]->isConstantEncoding()) {
     const A constA = args[0]->asUnchecked<ConstantVector<A>>()->valueAt(0);
     const A* __restrict rawA = &constA;
     const B constB = args[1]->asUnchecked<ConstantVector<B>>()->valueAt(0);
     const B* __restrict rawB = &constB;
     detail::applyAutoSimdComparisonInternal(
-        rows,
-        rawA,
-        rawB,
-        [&](const A* __restrict rawA, const B* __restrict rawB, int i) {
+        [&](const A* __restrict rawA,
+            const B* __restrict rawB,
+            const vector_size_t* __restrict indexA,
+            const vector_size_t* __restrict indexB,
+            int i) {
           if constexpr (sizeof...(cmpArgs) > 0) {
             return Compare::apply(rawA[0], rawB[0], cmpArgs...);
           } else {
             return cmp(rawA[0], rawB[0]);
           }
         },
-        result);
+        result,
+        rows,
+        rawA,
+        rawB,
+        nullptr,
+        nullptr);
+  } else if (
+      args[0]->isFlatEncoding() &&
+      args[1]->encoding() == VectorEncoding::Simple::DICTIONARY &&
+      args[1]
+          ->asUnchecked<DictionaryVector<B>>()
+          ->valueVector()
+          ->isFlatEncoding()) {
+    auto dictB = args[1]->asUnchecked<DictionaryVector<B>>();
+    const A* __restrict rawA =
+        args[0]->asUnchecked<FlatVector<A>>()->rawValues();
+    const B* __restrict rawB = dictB->valueVector()
+                                   ->template asUnchecked<FlatVector<B>>()
+                                   ->rawValues();
+    const vector_size_t* __restrict indexB =
+        dictB->indices()->template as<vector_size_t>();
+    detail::applyAutoSimdComparisonInternal(
+        [&](const A* __restrict rawA,
+            const B* __restrict rawB,
+            const vector_size_t* __restrict indexA,
+            const vector_size_t* __restrict indexB,
+            int i) {
+          if constexpr (sizeof...(cmpArgs) > 0) {
+            return Compare::apply(rawA[i], rawB[indexB[i]], cmpArgs...);
+          } else {
+            return cmp(rawA[i], rawB[indexB[i]]);
+          }
+        },
+        result,
+        rows,
+        rawA,
+        rawB,
+        nullptr,
+        indexB);
+  } else if (
+      args[1]->isFlatEncoding() &&
+      args[0]->encoding() == VectorEncoding::Simple::DICTIONARY &&
+      args[0]
+          ->asUnchecked<DictionaryVector<A>>()
+          ->valueVector()
+          ->isFlatEncoding()) {
+    auto dictA = args[0]->asUnchecked<DictionaryVector<A>>();
+    const A* __restrict rawA = dictA->valueVector()
+                                   ->template asUnchecked<FlatVector<A>>()
+                                   ->rawValues();
+    const vector_size_t* __restrict indexA =
+        dictA->indices()->template as<vector_size_t>();
+    const B* __restrict rawB =
+        args[1]->asUnchecked<FlatVector<B>>()->rawValues();
+    detail::applyAutoSimdComparisonInternal(
+        [&](const A* __restrict rawA,
+            const B* __restrict rawB,
+            const vector_size_t* __restrict indexA,
+            const vector_size_t* __restrict indexB,
+            int i) {
+          if constexpr (sizeof...(cmpArgs) > 0) {
+            return Compare::apply(rawA[indexA[i]], rawB[i], cmpArgs...);
+          } else {
+            return cmp(rawA[indexA[i]], rawB[i]);
+          }
+        },
+        result,
+        rows,
+        rawA,
+        rawB,
+        indexA,
+        nullptr);
   } else {
     VELOX_UNREACHABLE();
   }
+}
+
+template <typename A, typename B>
+bool shouldApplyAutoSimdComparison(
+    const SelectivityVector& rows,
+    std::vector<VectorPtr>& args) {
+  if (rows.end() - rows.begin() > 64) {
+    if ((args[0]->isFlatEncoding() || args[0]->isConstantEncoding()) &&
+        (args[1]->isFlatEncoding() || args[1]->isConstantEncoding())) {
+      return true;
+    } else if (
+        args[0]->isFlatEncoding() &&
+        args[1]->encoding() == VectorEncoding::Simple::DICTIONARY &&
+        args[1]
+            ->asUnchecked<DictionaryVector<B>>()
+            ->valueVector()
+            ->isFlatEncoding()) {
+      return true;
+    } else if (
+        args[1]->isFlatEncoding() &&
+        args[0]->encoding() == VectorEncoding::Simple::DICTIONARY &&
+        args[0]
+            ->asUnchecked<DictionaryVector<A>>()
+            ->valueVector()
+            ->isFlatEncoding()) {
+      return true;
+    }
+  }
+  return false;
 }
 } // namespace facebook::velox::functions