Support fast path for Spark comparison function in join filter
We observed the join filter's inputs are flat encoding(build side) and 
dictionary encoding(probe side). This PR extends the optimization we did in 
10273 to cover the input encoding: dict and flat.