@zhli1142015 2024-07-15
I added a benchmark to demonstrate the performance improvement this PR brings 
for comparison functions handling dict(flat) and flat inputs. The results are 
listed below. We also perf gain in Join of TPCDS Query72.


<html>

<head>

</head>

<body link="#467886" vlink="#96607D">


  | Perf Gain %
-- | --
Dict#BIGINT#0%null##equalto | 453%
Dict#BIGINT#0%null##greaterthan | 1279%
Dict#BIGINT#0%null##greaterthanorequal | 1271%
Dict#BIGINT#10%null##equalto | 183%
Dict#BIGINT#10%null##greaterthan | 312%
Dict#BIGINT#10%null##greaterthanorequal | 308%
Dict#DOUBLE#0%null##equalto | 243%
Dict#DOUBLE#0%null##greaterthan | 619%
Dict#DOUBLE#0%null##greaterthanorequal | 191%
Dict#DOUBLE#10%null##equalto | 171%
Dict#DOUBLE#10%null##greaterthan | 267%
Dict#DOUBLE#10%null##greaterthanorequal | 188%
Dict#INTEGER#0%null##equalto | 459%
Dict#INTEGER#0%null##greaterthan | 1438%
Dict#INTEGER#0%null##greaterthanorequal | 1197%
Dict#INTEGER#10%null##equalto | 180%
Dict#INTEGER#10%null##greaterthan | 297%
Dict#INTEGER#10%null##greaterthanorequal | 295%
Dict#REAL#0%null##equalto | 277%
Dict#REAL#0%null##greaterthan | 630%
Dict#REAL#0%null##greaterthanorequal | 184%
Dict#REAL#10%null##equalto | 158%
Dict#REAL#10%null##greaterthan | 263%
Dict#REAL#10%null##greaterthanorequal | 179%
Dict#SMALLINT#0%null##equalto | 430%
Dict#SMALLINT#0%null##greaterthan | 1270%
Dict#SMALLINT#0%null##greaterthanorequal | 1154%
Dict#SMALLINT#10%null##equalto | 173%
Dict#SMALLINT#10%null##greaterthan | 276%
Dict#SMALLINT#10%null##greaterthanorequal | 282%
Dict#TIMESTAMP#0%null##equalto | 404%
Dict#TIMESTAMP#0%null##greaterthan | 167%
Dict#TIMESTAMP#0%null##greaterthanorequal | 416%
Dict#TIMESTAMP#10%null##equalto | 173%
Dict#TIMESTAMP#10%null##greaterthan | 124%
Dict#TIMESTAMP#10%null##greaterthanorequal | 196%
Dict#TINYINT#0%null##equalto | 446%
Dict#TINYINT#0%null##greaterthan | 1151%
Dict#TINYINT#0%null##greaterthanorequal | 1085%
Dict#TINYINT#10%null##equalto | 170%
Dict#TINYINT#10%null##greaterthan | 277%
Dict#TINYINT#10%null##greaterthanorequal | 273%
Dict#VARCHAR#0%null##equalto | 393%
Dict#VARCHAR#0%null##greaterthan | 436%
Dict#VARCHAR#0%null##greaterthanorequal | 110%
Dict#VARCHAR#10%null##equalto | 168%
Dict#VARCHAR#10%null##greaterthan | 246%
Dict#VARCHAR#10%null##greaterthanorequal | 108%



</body>

</html>

Before
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
Dict#BIGINT#0%null##equalto                                 2.10ms    476.41
Dict#BIGINT#0%null##greaterthan                             5.24ms    190.67
Dict#BIGINT#0%null##greaterthanorequal                      5.21ms    191.83
----------------------------------------------------------------------------
Dict#BIGINT#10%null##equalto                                5.09ms    196.31
Dict#BIGINT#10%null##greaterthan                            7.34ms    136.19
Dict#BIGINT#10%null##greaterthanorequal                     7.27ms    137.62
----------------------------------------------------------------------------
Dict#BIGINT#90%null##equalto                              888.47us     1.13K
Dict#BIGINT#90%null##greaterthan                          907.16us     1.10K
Dict#BIGINT#90%null##greaterthanorequal                   892.12us     1.12K
----------------------------------------------------------------------------
Dict#DOUBLE#0%null##equalto                                 2.16ms    463.22
Dict#DOUBLE#0%null##greaterthan                             5.54ms    180.56
Dict#DOUBLE#0%null##greaterthanorequal                      5.58ms    179.35
----------------------------------------------------------------------------
Dict#DOUBLE#10%null##equalto                                5.20ms    192.40
Dict#DOUBLE#10%null##greaterthan                            7.60ms    131.62
Dict#DOUBLE#10%null##greaterthanorequal                     8.63ms    115.93
----------------------------------------------------------------------------
Dict#DOUBLE#90%null##equalto                              980.79us     1.02K
Dict#DOUBLE#90%null##greaterthan                          960.65us     1.04K
Dict#DOUBLE#90%null##greaterthanorequal                   970.52us     1.03K
----------------------------------------------------------------------------
Dict#INTEGER#0%null##equalto                                2.07ms    483.13
Dict#INTEGER#0%null##greaterthan                            5.69ms    175.63
Dict#INTEGER#0%null##greaterthanorequal                     4.80ms    208.28
----------------------------------------------------------------------------
Dict#INTEGER#10%null##equalto                               5.02ms    199.01
Dict#INTEGER#10%null##greaterthan                           7.10ms    140.88
Dict#INTEGER#10%null##greaterthanorequal                    7.03ms    142.31
----------------------------------------------------------------------------
Dict#INTEGER#90%null##equalto                             977.55us     1.02K
Dict#INTEGER#90%null##greaterthan                         976.95us     1.02K
Dict#INTEGER#90%null##greaterthanorequal                  988.36us     1.01K
----------------------------------------------------------------------------
Dict#REAL#0%null##equalto                                   2.11ms    472.96
Dict#REAL#0%null##greaterthan                               5.33ms    187.50
Dict#REAL#0%null##greaterthanorequal                        5.62ms    177.99
----------------------------------------------------------------------------
Dict#REAL#10%null##equalto                                  5.15ms    194.03
Dict#REAL#10%null##greaterthan                              7.44ms    134.33
Dict#REAL#10%null##greaterthanorequal                       8.39ms    119.18
----------------------------------------------------------------------------
Dict#REAL#90%null##equalto                                945.57us     1.06K
Dict#REAL#90%null##greaterthan                            949.77us     1.05K
Dict#REAL#90%null##greaterthanorequal                     934.47us     1.07K
----------------------------------------------------------------------------
Dict#SMALLINT#0%null##equalto                               1.96ms    510.77
Dict#SMALLINT#0%null##greaterthan                           5.07ms    197.38
Dict#SMALLINT#0%null##greaterthanorequal                    4.64ms    215.56
----------------------------------------------------------------------------
Dict#SMALLINT#10%null##equalto                              4.80ms    208.27
Dict#SMALLINT#10%null##greaterthan                          6.65ms    150.33
Dict#SMALLINT#10%null##greaterthanorequal                   6.69ms    149.48
----------------------------------------------------------------------------
Dict#SMALLINT#90%null##equalto                            937.32us     1.07K
Dict#SMALLINT#90%null##greaterthan                        920.06us     1.09K
Dict#SMALLINT#90%null##greaterthanorequal                 928.68us     1.08K
----------------------------------------------------------------------------
Dict#TIMESTAMP#0%null##equalto                              2.27ms    440.17
Dict#TIMESTAMP#0%null##greaterthan                          5.88ms    170.03
Dict#TIMESTAMP#0%null##greaterthanorequal                   7.12ms    140.52
----------------------------------------------------------------------------
Dict#TIMESTAMP#10%null##equalto                             5.08ms    196.69
Dict#TIMESTAMP#10%null##greaterthan                         8.44ms    118.54
Dict#TIMESTAMP#10%null##greaterthanorequal                  9.25ms    108.12
----------------------------------------------------------------------------
Dict#TIMESTAMP#90%null##equalto                           887.51us     1.13K
Dict#TIMESTAMP#90%null##greaterthan                       885.68us     1.13K
Dict#TIMESTAMP#90%null##greaterthanorequal                882.42us     1.13K
----------------------------------------------------------------------------
Dict#TINYINT#0%null##equalto                                2.02ms    495.50
Dict#TINYINT#0%null##greaterthan                            4.63ms    216.08
Dict#TINYINT#0%null##greaterthanorequal                     4.62ms    216.28
----------------------------------------------------------------------------
Dict#TINYINT#10%null##equalto                               4.81ms    207.69
Dict#TINYINT#10%null##greaterthan                           6.63ms    150.91
Dict#TINYINT#10%null##greaterthanorequal                    6.56ms    152.33
----------------------------------------------------------------------------
Dict#TINYINT#90%null##equalto                             941.47us     1.06K
Dict#TINYINT#90%null##greaterthan                         921.85us     1.08K
Dict#TINYINT#90%null##greaterthanorequal                  916.04us     1.09K
----------------------------------------------------------------------------
Dict#VARCHAR#0%null##equalto                                2.22ms    451.19
Dict#VARCHAR#0%null##greaterthan                           11.21ms     89.24
Dict#VARCHAR#0%null##greaterthanorequal                    12.09ms     82.72
----------------------------------------------------------------------------
Dict#VARCHAR#10%null##equalto                               5.01ms    199.76
Dict#VARCHAR#10%null##greaterthan                          10.96ms     91.24
Dict#VARCHAR#10%null##greaterthanorequal                   12.39ms     80.68
----------------------------------------------------------------------------
Dict#VARCHAR#90%null##equalto                             846.83us     1.18K
Dict#VARCHAR#90%null##greaterthan                         840.21us     1.19K
Dict#VARCHAR#90%null##greaterthanorequal                  862.67us     1.16K
```
After
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
Dict#BIGINT#0%null##equalto                               381.44us     2.62K
Dict#BIGINT#0%null##greaterthan                           381.02us     2.62K
Dict#BIGINT#0%null##greaterthanorequal                    380.73us     2.63K
----------------------------------------------------------------------------
Dict#BIGINT#10%null##equalto                                1.80ms    556.79
Dict#BIGINT#10%null##greaterthan                            1.78ms    563.18
Dict#BIGINT#10%null##greaterthanorequal                     1.78ms    560.96
----------------------------------------------------------------------------
Dict#BIGINT#90%null##equalto                              765.27us     1.31K
Dict#BIGINT#90%null##greaterthan                          808.06us     1.24K
Dict#BIGINT#90%null##greaterthanorequal                   730.84us     1.37K
----------------------------------------------------------------------------
Dict#DOUBLE#0%null##equalto                               637.44us     1.57K
Dict#DOUBLE#0%null##greaterthan                           776.43us     1.29K
Dict#DOUBLE#0%null##greaterthanorequal                      1.92ms    520.37
----------------------------------------------------------------------------
Dict#DOUBLE#10%null##equalto                                1.92ms    520.30
Dict#DOUBLE#10%null##greaterthan                            2.07ms    483.57
Dict#DOUBLE#10%null##greaterthanorequal                     3.00ms    332.90
----------------------------------------------------------------------------
Dict#DOUBLE#90%null##equalto                              821.65us     1.22K
Dict#DOUBLE#90%null##greaterthan                          797.34us     1.25K
Dict#DOUBLE#90%null##greaterthanorequal                   875.24us     1.14K
----------------------------------------------------------------------------
Dict#INTEGER#0%null##equalto                              372.72us     2.68K
Dict#INTEGER#0%null##greaterthan                          374.38us     2.67K
Dict#INTEGER#0%null##greaterthanorequal                   373.05us     2.68K
----------------------------------------------------------------------------
Dict#INTEGER#10%null##equalto                               1.79ms    559.27
Dict#INTEGER#10%null##greaterthan                           1.79ms    559.70
Dict#INTEGER#10%null##greaterthanorequal                    1.78ms    562.68
----------------------------------------------------------------------------
Dict#INTEGER#90%null##equalto                             814.92us     1.23K
Dict#INTEGER#90%null##greaterthan                         841.07us     1.19K
Dict#INTEGER#90%null##greaterthanorequal                  855.65us     1.17K
----------------------------------------------------------------------------
Dict#REAL#0%null##equalto                                 562.68us     1.78K
Dict#REAL#0%null##greaterthan                             739.73us     1.35K
Dict#REAL#0%null##greaterthanorequal                        1.98ms    505.03
----------------------------------------------------------------------------
Dict#REAL#10%null##equalto                                  2.00ms    499.09
Dict#REAL#10%null##greaterthan                              2.05ms    487.97
Dict#REAL#10%null##greaterthanorequal                       3.01ms    332.12
----------------------------------------------------------------------------
Dict#REAL#90%null##equalto                                896.75us     1.12K
Dict#REAL#90%null##greaterthan                            836.78us     1.20K
Dict#REAL#90%null##greaterthanorequal                     868.67us     1.15K
----------------------------------------------------------------------------
Dict#SMALLINT#0%null##equalto                             373.97us     2.67K
Dict#SMALLINT#0%null##greaterthan                         371.69us     2.69K
Dict#SMALLINT#0%null##greaterthanorequal                  370.91us     2.70K
----------------------------------------------------------------------------
Dict#SMALLINT#10%null##equalto                              1.76ms    566.68
Dict#SMALLINT#10%null##greaterthan                          1.77ms    564.70
Dict#SMALLINT#10%null##greaterthanorequal                   1.75ms    570.12
----------------------------------------------------------------------------
Dict#SMALLINT#90%null##equalto                            755.88us     1.32K
Dict#SMALLINT#90%null##greaterthan                        778.25us     1.28K
Dict#SMALLINT#90%null##greaterthanorequal                 759.63us     1.32K
----------------------------------------------------------------------------
Dict#TIMESTAMP#0%null##equalto                            451.09us     2.22K
Dict#TIMESTAMP#0%null##greaterthan                          2.20ms    454.34
Dict#TIMESTAMP#0%null##greaterthanorequal                   1.38ms    727.21
----------------------------------------------------------------------------
Dict#TIMESTAMP#10%null##equalto                             1.86ms    537.48
Dict#TIMESTAMP#10%null##greaterthan                         3.76ms    265.83
Dict#TIMESTAMP#10%null##greaterthanorequal                  3.13ms    319.55
----------------------------------------------------------------------------
Dict#TIMESTAMP#90%null##equalto                           782.17us     1.28K
Dict#TIMESTAMP#90%null##greaterthan                       825.13us     1.21K
Dict#TIMESTAMP#90%null##greaterthanorequal                755.28us     1.32K
----------------------------------------------------------------------------
Dict#TINYINT#0%null##equalto                              374.32us     2.67K
Dict#TINYINT#0%null##greaterthan                          371.18us     2.69K
Dict#TINYINT#0%null##greaterthanorequal                   396.36us     2.52K
----------------------------------------------------------------------------
Dict#TINYINT#10%null##equalto                               1.78ms    561.88
Dict#TINYINT#10%null##greaterthan                           1.76ms    568.81
Dict#TINYINT#10%null##greaterthanorequal                    1.76ms    567.13
----------------------------------------------------------------------------
Dict#TINYINT#90%null##equalto                             868.68us     1.15K
Dict#TINYINT#90%null##greaterthan                         867.86us     1.15K
Dict#TINYINT#90%null##greaterthanorequal                  852.31us     1.17K
----------------------------------------------------------------------------
Dict#VARCHAR#0%null##equalto                              452.35us     2.21K
Dict#VARCHAR#0%null##greaterthan                            2.09ms    477.34
Dict#VARCHAR#0%null##greaterthanorequal                     5.76ms    173.70
----------------------------------------------------------------------------
Dict#VARCHAR#10%null##equalto                               1.87ms    535.40
Dict#VARCHAR#10%null##greaterthan                           3.17ms    315.94
Dict#VARCHAR#10%null##greaterthanorequal                    5.96ms    167.69
----------------------------------------------------------------------------
Dict#VARCHAR#90%null##equalto                             766.55us     1.30K
Dict#VARCHAR#90%null##greaterthan                         811.50us     1.23K
Dict#VARCHAR#90%null##greaterthanorequal                  752.79us     1.33K
----------------------------------------------------------------------------

```

Query 72 Join
```
(19) ShuffledHashJoinExecTransformer
Left keys [1]: [cs_item_sk#14]
Right keys [1]: [inv_item_sk#122]
Join type: Inner
Join condition: (inv_quantity_on_hand#124 < cs_quantity#17)
```
Before
![image](https://github.com/user-attachments/assets/dff15088-84c8-4358-ac71-48cff0200119)

After
![image](https://github.com/user-attachments/assets/a24c829d-e567-4cf0-89ed-c352f7ef88d8)